Issues #4, #19, #23, #25
Issue #4 The Header and footer of the Blog, now are in the Platform, it
last to adjust the gap in the footer, and some css adjustments, I ask to
SA for help with this.
Files: -index.php - index2.php -
Issue #19 Modification of the file climateAnalogues_v1.js
Issue #23 Modification of the file climateAnalogues_v1.js
Modify the InitTooltips()
Not finished, last the "Add Rotation to the Parameters".
Issue #25 Modification of the file climateAnalogues_v1.js
Modify the createSteps()  in step 2 to left only the options 30
arc-seconds, 2.5 and 10 arc-min